+++ ELITPCdcs/clientGUI/include/mks946widget.h
@@ -45,6 +45,7 @@ class MKS946Widget : public AbstractWidget {
   private:
     std::vector<DCSRelayWidget *> relayWidgets;
     bool connectionState;
+    bool PIDState;
     // create layout procedures and variables
     QVBoxLayout *mainLayout;
 
@@ -104,6 +105,8 @@ class MKS946Widget : public AbstractWidget {
 
     void fillGrid(std::vector<std::pair<std::string, QWidget *>> names, QGridLayout *grid,
                   size_t grid_max = 100);
+
+    std::string getDeviceType() const final { return "MKS946"; }
 };
 
 #endif  // MKS946WIDGET_H